moving tests to respective folders for maintainance
