+++ tests/Command/Branch/BranchChangelogCommandTest.php
@@ -9,9 +9,10 @@
  * with this source code in the file LICENSE.
  */
 
-namespace Gush\Tests\Command;
+namespace Gush\Tests\Command\Branch;
 
 use Gush\Command\Branch\BranchChangelogCommand;
+use Gush\Tests\Command\BaseTestCase;
 use Gush\Tests\Fixtures\OutputFixtures;
 
 class BranchChangelogCommandTest extends BaseTestCase
