+++ tests/Command/Branch/BranchForkCommandTest.php
@@ -9,27 +9,26 @@
  * with this source code in the file LICENSE.
  */
 
-namespace Gush\Tests\Command;
+namespace Gush\Tests\Command\Branch;
 
-use Gush\Command\Branch\BranchPushCommand;
+use Gush\Command\Branch\BranchForkCommand;
+use Gush\Tests\Command\BaseTestCase;
 use Gush\Tests\Fixtures\OutputFixtures;
 
-class BranchPushCommandTest extends BaseTestCase
+class BranchForkCommandTest extends BaseTestCase
 {
     const TEST_USERNAME = 'cordoval';
 
     public function testCommand()
     {
         $processHelper = $this->expectProcessHelper();
-        $gitHelper = $this->expectGitHelper();
         $this->expectsConfig();
-        $tester = $this->getCommandTester($command = new BranchPushCommand());
+        $tester = $this->getCommandTester($command = new BranchForkCommand());
         $command->getHelperSet()->set($processHelper, 'process');
-        $command->getHelperSet()->set($gitHelper, 'git');
 
-        $tester->execute(['--org' => 'cordoval', '--repo' => 'gush'], ['interactive' => false]);
+        $tester->execute(['--org' => 'gushphp', '--repo' => 'gush'], ['interactive' => false]);
 
-        $this->assertEquals(OutputFixtures::BRANCH_PUSH, trim($tester->getDisplay(true)));
+        $this->assertEquals(OutputFixtures::BRANCH_FORK, trim($tester->getDisplay(true)));
     }
 
     private function expectProcessHelper()
@@ -42,30 +41,13 @@ class BranchPushCommandTest extends BaseTestCase
             ->method('runCommands')
             ->with([
                 [
-                    'line' => 'git push -u cordoval some-branch',
+                    'line' => 'git remote add cordoval git@github.com:cordoval/gush.git',
                     'allow_failures' => true
                 ],
             ])
         ;
 
         return $processHelper;
-    }
-
-    private function expectGitHelper()
-    {
-        $gitHelper = $this
-            ->getMockBuilder('Gush\Helper\GitHelper')
-            ->disableOriginalConstructor()
-            ->setMethods(['getBranchName'])
-            ->getMock()
-        ;
-        $gitHelper
-            ->expects($this->once())
-            ->method('getBranchName')
-            ->will($this->returnValue('some-branch'))
-        ;
-
-        return $gitHelper;
     }
 
     private function expectsConfig()
