--- tests/Command/IssueMilestoneListCommandTest.php
@@ -1,25 +1,0 @@
-<?php
-
-/**
- * This file is part of Gush package.
- *
- * (c) 2013-2014 Luis Cordova <cordoval@gmail.com>
- *
- * This source file is subject to the MIT license that is bundled
- * with this source code in the file LICENSE.
- */
-
-namespace Gush\Tests\Command;
-
-use Gush\Command\Issue\IssueMilestoneListCommand;
-
-class IssueMilestoneListCommandTest extends BaseTestCase
-{
-    public function testCommand()
-    {
-        $tester = $this->getCommandTester(new IssueMilestoneListCommand());
-        $tester->execute(['--org' => 'gushphp', '--repo' => 'gush'], ['interactive' => false]);
-
-        $this->assertEquals("version 1.0", trim($tester->getDisplay(true)));
-    }
-}
