+++ tests/Command/PullRequest/PullrequestAssignCommandTest.php
@@ -9,10 +9,11 @@
  * with this source code in the file LICENSE.
  */
 
-namespace Gush\Tests\Command;
+namespace Gush\Tests\Command\PullRequest;
 
 use Gush\Command\PullRequest\PullRequestAssignCommand;
 use Gush\Tester\Adapter\TestAdapter;
+use Gush\Tests\Command\BaseTestCase;
 
 class PullRequestAssignCommandTest extends BaseTestCase
 {
