+++ tests/Command/Util/MetaHeaderCommandTest.php
@@ -9,11 +9,12 @@
  * with this source code in the file LICENSE.
  */
 
-namespace Gush\Tests\Command;
+namespace Gush\Tests\Command\Util;
 
 use Gush\Command\Util\MetaHeaderCommand;
 use Gush\Helper\MetaHelper;
 use Gush\Meta as Meta;
+use Gush\Tests\Command\BaseTestCase;
 use Gush\Tests\Fixtures\OutputFixtures;
 
 class MetaHeaderCommandTest extends BaseTestCase
