+++ tests/Command/Util/VersionEyeCommandTest.php
@@ -9,9 +9,10 @@
  * with this source code in the file LICENSE.
  */
 
-namespace Gush\Tests\Command;
+namespace Gush\Tests\Command\Util;
 
 use Gush\Command\Util\VersionEyeCommand;
+use Gush\Tests\Command\BaseTestCase;
 use Gush\Tests\Fixtures\OutputFixtures;
 
 class VersionEyeCommandTest extends BaseTestCase
